+++ frontend/src/views/LandingPage.vue
@@ -4,69 +4,22 @@
       You're logged in!
     </div>
     <div v-else>
-      <h3>Login!</h3>
-      <CForm @submit="onLogin">
-        <div>
-          <CFormInput
-            type="email"
-            id="exampleFormControlInput1"
-            label="Email"
-            placeholder="name@example.com"
-            aria-describedby="exampleFormControlInputHelpInline"
-            v-model="loginEmail"
-          />
-        </div>
-        <div>
-          <CFormLabel for="inputPassword2">Password</CFormLabel>
-          <CFormInput type="password" id="inputPassword2" placeholder="Password" v-model="loginPassword" />
-        </div>
-        <div class="button-container">
-          <CButton type="submit" color="primary" value="Login">Login</CButton>
-        </div>
-      </CForm>
+      <LoginFormVue />
     </div>
   </div>
 </template>
 
 <script>
   import "@/store/index";
-  import { mapActions, mapGetters } from "vuex";
+  import { mapGetters } from "vuex";
+  import LoginFormVue from "@/components/LoginForm.vue";
   export default {
     name: "Landing Page",
+    components: {
+      LoginFormVue
+    },
     computed: {
         ...mapGetters(["isLoggedIn"]),
-    },
-    data() {
-      return {
-        signUpEmail: "",
-        signUpPassword: "",
-        loginEmail: "",
-        loginPassword: "",
-        errors: "",
-      };
-    },
-    methods: {
-      ...mapActions(["loginUser", "logoutUser"]),
-      onLogin(event) {
-        // Reset the alert to empty
-        this.$store.state.errors = "";
-        event.preventDefault();
-        let data = {
-            user: {
-                email: this.loginEmail,
-                password: this.loginPassword,
-            },
-        };
-        this.loginUser(data)
-        this.resetData();
-      },
-      resetData() {
-        this.signUpEmail = "";
-        this.signUpPassword = "";
-        this.loginEmail = "";
-        this.loginPassword = "";
-      },
     }
-  }
-
+  };
 </script>
